Struct udev::Monitor

pub struct Monitor<'u> {
    pub udev: &'u Udev,
    // some fields omitted
}

Fields

udev

Methods

impl<'u> Monitor<'u>

fn filter_by_subsystem(self, subsystem: &str) -> Monitor<'u>

Filter by subsystem.

Exclude devices that don't match the specified subsystem or a previously specified subsystem.

fn filter_by_subsystem_devtype(self, subsystem: &str, devtype: &str) -> Monitor<'u>

Filter by subsystem/devtype combination.

Exclude devices that don't match the specified subsystem/devtype combination or a previously specified subsystem/devtype combination (or any subsystem previously specified in a filter_subsystem invocation).

fn filter_by_tag(self, tag: &str) -> Monitor<'u>

Filter by tag.

Exclude devices that don't match the specified tag or a previously specified tag.

fn clear_filters(self) -> Monitor<'u>

Reset all filters on this monitor. No devices will be excluded.

fn iter<'m>(&'m self) -> MonitorIterator<'m, 'u>

Iterate over udev events.

The returned iterator will block on calls to next until their a device is available.

The returned iterator will never end (next will never return None).