udev

Struct MonitorSocket

pub struct MonitorSocket { /* private fields */ }

An active monitor that can receive events.

The events received by a Socket match the filters set up by the Monitor that created the socket.

Monitors are initially set up to receive events from the kernel via a nonblocking socket. A variant of poll() should be used on the file descriptor returned by the AsRawFd trait to wait for new events.

Implementations

impl Socket

pub fn iter(&self) -> SocketIter<'_>

Create an iterator of socket event messages