udev
Trait FromRaw

pub trait FromRaw<T: 'static> {
    // Required method
    unsafe fn from_raw(ptr: *mut T) -> Self;
}

Convert from a raw pointer

Required Methods

unsafe fn from_raw(ptr: *mut T) -> Self

Create an object from a given raw pointer.

The reference count will not be increased, be sure not to free this pointer.

Safety

The pointer has to be a valid reference to the expected underlying udev-struct or undefined behaviour might occur.

Object Safety

This trait is not object safe.