Crate udev

libudev Bindings for Rust

Re-exports

pub extern crate libudev_sys as ffi;

Structs

Device

A structure that provides access to sysfs/kernel devices.

Entry

Rust wrapper for each entry in List, each of which contains a name and a value.

Enumerator

An enumeration context.

Event

An event that indicates a change in device state.

List

Rust wrapper for the udev_list_entry struct, which provides sequential access to an associative list of string names and values.

MonitorBuilder

Monitors for device events.

MonitorSocket

An active monitor that can receive events.

Udev

Rust wrapper for the udev struct which represents an opaque libudev context

Enums

EventType

Types of events that can be received from udev.

Traits

AsRaw

Receive the underlying raw pointer

AsRawWithContext

Receive the underlying raw pointer for types with an associated udev struct which must outlive them.

FromRaw

Convert from a raw pointer

FromRawWithContext

Convert from a raw pointer for types which must be associated with a Udev context object.

Type Definitions

Attributes

A convenience alias for a list of attributes, bound to a device.

Devices

Iterator over devices.

Properties

A convenience alias for a list of properties, bound to a device.