udev

Struct MonitorBuilder

pub struct MonitorBuilder { /* private fields */ }

Monitors for device events.

A monitor communicates with the kernel over a socket. Filtering events is performed efficiently in the kernel, and only events that match the filters are received by the socket. Filters must be set up before listening for events.

Implementations

impl Builder

pub fn new() -> Result<Self>

Creates a new udev event Monitor.

pub fn new_kernel() -> Result<Self>

Creates a new kernel event Monitor.

pub fn match_subsystem<T: AsRef<OsStr>>(self, subsystem: T) -> Result<Self>

Adds a filter that matches events for devices with the given subsystem.

pub fn match_subsystem_devtype<T: AsRef<OsStr>, U: AsRef<OsStr>>( self, subsystem: T, devtype: U, ) -> Result<Self>

Adds a filter that matches events for devices with the given subsystem and device type.

pub fn match_tag<T: AsRef<OsStr>>(self, tag: T) -> Result<Self>

Adds a filter that matches events for devices with the given tag.

pub fn clear_filters(self) -> Result<Self>

Removes all filters currently set on the monitor.

pub fn listen(self) -> Result<Socket>

Listens for events matching the current filters.

This method consumes the Monitor.